+++ src/fourchess/pieces/Bishop.groovy
@@ -18,12 +18,12 @@ public class Bishop extends Piece
 	public Set<Point> getValidMoves(Point position)
 	{
 		Set<Point> moves = super.getValidMoves(position);
-		Piece piece = board.getPiece(position.x, position.y);
+		Piece piece = board.getPiece(position.x.intValue(), position.y.intValue());
 		
-		Rules.checkPath(piece, position.x, position.y, -1, -1, moves); // Up-Left
-		Rules.checkPath(piece, position.x, position.y, -1, 1, moves);  // Down-Left
-		Rules.checkPath(piece, position.x, position.y, 1, -1, moves);  // Up-Right 
-		Rules.checkPath(piece, position.x, position.y, 1, 1, moves);   // Down-Right
+		Rules.checkPath(piece, position.x.intValue(), position.y.intValue(), -1, -1, moves); // Up-Left
+		Rules.checkPath(piece, position.x.intValue(), position.y.intValue(), -1, 1, moves);  // Down-Left
+		Rules.checkPath(piece, position.x.intValue(), position.y.intValue(), 1, -1, moves);  // Up-Right
+		Rules.checkPath(piece, position.x.intValue(), position.y.intValue(), 1, 1, moves);   // Down-Right
 		
 		return moves;
 	}
